fix(alertsmanagement): update AlertSeverity type from Long to Float
- Change AlertSeverity implementation to use Float instead of Long
- Update constructor, getValue(), fromValue(), toJson() methods to handle Float values
- Modify deserialization in LogAlertRuleProperties to read severity as float
- Update static constants to use Float literals (0f, 1f, etc.)

fix(alertsmanagement): correct API path case sensitivity

- Fix provider name from "Microsoft.AlertsManagement" to "microsoft.AlertsManagement"
- Update both async and sync service method annotations

refactor(alertsmanagement): standardize parameter order in service calls

- Reorder method parameters to place apiVersion before resourceId
- Update both async and sync method invocations consistently

docs(alertsmanagement): improve parameter documentation

- Replace generic descriptions with more specific parameter names
- Change "The identifier of the resource" to "The resourceId parameter"
- Change "The alert rule to preview" to "The parameters parameter"

+++ sdk/alertsmanagement/azure-resourcemanager-previewalertrule/src/main/java/com/azure/app/fluent/models/LogAlertRuleProperties.java
@@ -353,7 +353,7 @@ public final class LogAlertRuleProperties implements JsonSerializable<LogAlertRu
                 } else if ("displayName".equals(fieldName)) {
                     deserializedLogAlertRuleProperties.displayName = reader.getString();
                 } else if ("severity".equals(fieldName)) {
-                    deserializedLogAlertRuleProperties.severity = AlertSeverity.fromValue(reader.getLong());
+                    deserializedLogAlertRuleProperties.severity = AlertSeverity.fromValue(reader.getFloat());
                 } else if ("enabled".equals(fieldName)) {
                     deserializedLogAlertRuleProperties.enabled = reader.getNullable(JsonReader::getBoolean);
                 } else if ("scopes".equals(fieldName)) {
